# openapi.yaml
openapi: '3.0.0'
info:
  title: whoami
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
tags:
  - name: app
    description: Information about the app instance
  - name: environment
    description: Information about the run environment
  - name: stress
    description: Stress tests for hardware resources
  - name: client
    description: Information about the client
components:
  schemas:
    version:
      type: object
      properties:
        version:
          type: string
        major:
          type: integer
          format: int32
          minimum: 0
        minor:
          type: integer
          format: int32
          minimum: 0
        patch:
          type: integer
          format: int32
          minimum: 0
        commit:
          type: string
        buildDate:
          type: string
          format: date-time
    memory:
      type: object
      properties:
        total:
          type: integer
          format: int64
          minimum: 0
        available:
          type: integer
          format: int64
          minimum: 0
        used:
          type: integer
          format: int64
          minimum: 0
        usedPercent:
          type: number
          format: double
          minimum: 0
          maximum: 100
    interface:
      type: object
      properties:
        index:
          type: integer
          format: int64
          minimum: 1
        mtu:
          type: integer
          format: int64
          minimum: 0
        name:
          type: string
        addresses:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: string
    hostnameInfo:
      type: object
      properties:
        hostname:
          type: string
    memoryStressSession:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        bytesAllocated:
          type: integer
          format: int64
    memoryStressSessionForm:
      type: object
      properties:
        interval:
          type: integer
          format: duration
          minimum: 0
        allocSize:
          type: integer
          format: int64
          minimum: 0
    request:
      type: object
      properties:
        method:
          type: string
        path:
          type: string
        proto:
          type: string
        protoMajor:
          type: integer
          format: int32
          minimum: 0
        protoMinor:
          type: integer
          format: int32
          minimum: 0
        headers:
          type: object
        form:
          type: object
        trailerHeaders:
          type: object
        host:
          type: string
        remoteAddress:
          type: string
    requestForm:
      type: object
paths:
  /version:
    get:
      tags:
        - app
      summary: Returns the running version
      responses:
        '200':
          description: A JSON object describing the app version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/version'
  /memory:
    get:
      tags:
        - environment
      summary: Returns the available RAM
      responses:
        '200':
          description: A JSON object describing the available RAM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/memory'
  /interfaces:
    get:
      tags:
        - environment
      summary: Returns the network interfaces
      responses:
        '200':
          description: A JSON object describing the network interfaces as seen by the running environment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/interface'
  /hostname:
      get:
        tags:
          - environment
        summary: Returns the machine's hostname
        responses:
          '200':
            description: A JSON object describing the machine's hostname
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/interface'
  /memory/stresssession:
    get:
      tags:
        - stress
      summary: Get a memory stress session
      responses:
        '200':
          description: Info about the current memory stress session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/memoryStressSession'
    post:
      tags:
        - stress
      summary: Start a memory stress session
      requestBody:
        description: Memory stress parameters
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/memoryStressSessionForm'
      responses:
        '200':
          description: Info about the new memory stress session that has just started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/memoryStressSession'
    delete:
      tags:
        - stress
      summary: Cancel a memory stress session
      responses:
        '200':
          description: Info about the memory stress session that has just been cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/memoryStressSession'
  /request:
    get:
      tags:
        - client
      summary: Echo a GET request
      responses:
        '200':
          description: Request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/request'
    post:
      tags:
        - client
      summary: Echo a POST request
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/requestForm'
      responses:
        '200':
          description: Request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/request'
    put:
      tags:
        - client
      summary: Echo a PUT request
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/requestForm'
      responses:
        '200':
          description: Request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/request'
    delete:
      tags:
        - client
      summary: Echo a DELETE request
      responses:
        '200':
          description: Request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/request'
    options:
      tags:
        - client
      summary: Echo a OPTIONS request
      responses:
        '200':
          description: Request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/request'
    trace:
      tags:
        - client
      summary: Echo a TRACE request
      responses:
        '200':
          description: Request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/request'
    patch:
        tags:
          - client
        summary: Echo a PATCH request
        requestBody:
          content:
            application/x-www-form-urlencoded:
              schema:
                $ref: '#/components/schemas/requestForm'
        responses:
          '200':
            description: Request details
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/request'